+++ src/virtualOffset.ts
@@ -11,24 +11,27 @@ export default class VirtualOffset {
   }
 
   compareTo(b: VirtualOffset) {
-    return (
-      this.blockPosition - b.blockPosition || this.dataPosition - b.dataPosition
-    )
+    return this.blockPosition - b.blockPosition || this.dataPosition - b.dataPosition
   }
 
   static min(...args: VirtualOffset[]) {
     let min
     let i = 0
-    for (; !min; i += 1) min = args[i]
+    for (; !min; i += 1) {
+      min = args[i]
+    }
     for (; i < args.length; i += 1) {
-      if (min.compareTo(args[i]) > 0) min = args[i]
+      if (min.compareTo(args[i]) > 0) {
+        min = args[i]
+      }
     }
     return min
   }
 }
 export function fromBytes(bytes: Buffer, offset = 0, bigendian = false) {
-  if (bigendian)
+  if (bigendian) {
     throw new Error('big-endian virtual file offsets not implemented')
+  }
 
   return new VirtualOffset(
     bytes[offset + 7] * 0x10000000000 +