+++ test/bai.test.js
@@ -161,7 +161,7 @@ describe('1000 genomes bam check', () => {
       csiPath: require.resolve('./data/1000genomes_hg00096_chr1.bam.csi'),
     })
     await ti.getHeader()
-    const records = await ti.getRecordsForRange('1', 0, 1000)
+    const records = await ti.getRecordsForRange('1', 0, 10000)
     expect(records).toMatchSnapshot()
   })
   it('deep check 1000 genomes bai', async () => {
@@ -169,7 +169,7 @@ describe('1000 genomes bam check', () => {
       bamPath: require.resolve('./data/1000genomes_hg00096_chr1.bam'),
     })
     await ti.getHeader()
-    const records = await ti.getRecordsForRange('1', 0, 1000)
+    const records = await ti.getRecordsForRange('1', 0, 10000)
     expect(records).toMatchSnapshot()
   })
   it('start to deep check 1000 genomes but abort instead', async () => {
@@ -485,6 +485,30 @@ test('long read consistent IDs chm1 pacbio', async () => {
   expect(r1.id()).toEqual(r2.id())
 })
 
+test('long read consistent IDs hg002 nanopore', async () => {
+  const ti = new BamFile({
+    bamPath: require.resolve('./data/out.bam'),
+  })
+  await ti.getHeader()
+  const ret1 = await ti.getRecordsForRange('1', 200000, 700000)
+  const ret2 = await ti.getRecordsForRange('1', 700000, 900000)
+  const findfeat = k => k.get('name') === '901e456a-b7cf-4624-bd27-e2981a6c7ca5'
+  const [r1, r2] = [ret1, ret2].map(x => x.find(findfeat))
+  expect(r1.id()).toEqual(r2.id())
+})
+
+test('long read consistent IDs hg002 nanopore 2', async () => {
+  const ti = new BamFile({
+    bamPath: require.resolve('./data/out.bam'),
+  })
+  await ti.getHeader()
+  const ret3 = await ti.getRecordsForRange('1', 200000, 560000)
+  const ret4 = await ti.getRecordsForRange('1', 500000, 700000)
+  const findfeat2 = k => k.get('name') === '7c318bae-cae9-4cf9-8efc-c761304aa0da'
+  const [r3, r4] = [ret3, ret4].map(x => x.find(findfeat2))
+  expect(r3.id()).toEqual(r4.id())
+})
+
 test('long tag list', async () => {
   const ti = new BamFile({
     bamPath: require.resolve('./data/long_tag_list.bam'),
@@ -495,28 +519,32 @@ test('long tag list', async () => {
   expect(ret1[0]._tags()).toMatchSnapshot()
 })
 
-test('fix decoding error on ID', async () => {
+test('fix decoding error for ID tag', async () => {
   const ti1 = new BamFile({
     bamPath: require.resolve('./data/long_tag_list.bam'),
   })
-  const ti2 = new BamFile({
-    bamPath: require.resolve('./data/long_tag_list2.bam'),
-  })
+
   await ti1.getHeader()
-  await ti2.getHeader()
   const ret1 = await ti1.getRecordsForRange('1', 0, 3000000)
-  const ret2 = await ti2.getRecordsForRange('1', 0, 3000000)
-  expect(ret2[0].get('DI')).toBe(78190)
-  expect(ret2[1].get('DI')).toBe(4440)
   expect(ret1[0].get('ID')).toBe(78190)
   expect(ret1[1].get('ID')).toBe(4440)
 })
+test('fix decoding error for DI tag', async () => {
+  const ti1 = new BamFile({
+    bamPath: require.resolve('./data/long_tag_list2.bam'),
+  })
+  await ti1.getHeader()
+  const ret1 = await ti1.getRecordsForRange('1', 0, 3000000)
+  expect(ret1[0].get('DI')).toBe(78190)
+  expect(ret1[1].get('DI')).toBe(4440)
+})
+
 xtest('large chunks', async () => {
   const ti = new BAI({
     filehandle: new LocalFile(require.resolve('./data/out.marked.bai')),
   })
 
-  const ret = await ti.parse()
+  await ti.parse()
   //index 16 == 'chr17'
   const ret1 = await ti.blocksForRange(16, 41248671, 41337570)
   expect(ret1[0].fetchedSize()).toBe(10893136)