+++ test/csi.test.js
@@ -413,7 +413,7 @@ xtest('large chunks', async () => {
     filehandle: new LocalFile(require.resolve('./data/out.marked.csi')),
   })
 
-  const ret = await ti.parse()
+  await ti.parse()
   //index 16 == 'chr17'
   const ret1 = await ti.blocksForRange(16, 41248671, 41337570)
   expect(ret1[0].fetchedSize()).toBe(10893136)