+++ CHANGELOG.md
@@ -1,3 +1,17 @@
+<a name="1.0.36"></a>
+## [1.0.36](https://github.com/GMOD/bam-js/compare/v1.0.35...v1.0.36) (2020-03-05)
+
+
+
+- Adds a shortcut to stop parsing chunks after a record is detected to be outside the requested range while decoding
+
+<a name="1.0.35"></a>
+## [1.0.35](https://github.com/GMOD/bam-js/compare/v1.0.34...v1.0.35) (2020-02-04)
+
+
+
+- Update scheme used to calculate unique fileOffset based IDs using @gmod/bgzf-filehandle updates
+
 <a name="1.0.34"></a>
 ## [1.0.34](https://github.com/GMOD/bam-js/compare/v1.0.33...v1.0.34) (2020-01-24)